Create Finding duplicate values in String using Java
To find the duplicate values from a string first at first we need to convert it into char array as a sequence of character. Then we are using nested for loop along with a if condition to find the duplicate values. Output shows the duplicate values. The output will come as follows:
The string is: beautiful flower
Duplicate Characters in above string are: e u f l